+++ src/error.rs
@@ -1,5 +1,6 @@
 use std::env::VarError;
 use std::io::Error as IOError;
+use std::path::PathBuf;
 use std::str::Utf8Error;
 use std::string::FromUtf8Error;
 
@@ -14,45 +15,48 @@ quick_error! {
     pub enum FatalError {
         IOError(err: IOError) {
             from()
-            cause(err)
+            source(err)
             display("IO Error: {}", err)
+        }
+        FileNotFound(filename: PathBuf){
+            display("Unable to find file {} to perform replace", filename.display())
         }
         InvalidCargoFileFormat(err: TomlError) {
             display("Invalid TOML file format: {}", err)
             from()
-            cause(err)
+            source(err)
         }
         InvalidCargoFileFormat2(err: TomlEditError) {
             display("Invalid TOML file format: {}", err)
             from()
-            cause(err)
+            source(err)
         }
         InvalidCargoFileFormat3(err: CargoMetaError) {
             display("Invalid TOML file format: {}", err)
             from()
-            cause(err)
+            source(err)
         }
         InvalidCargoConfigKeys {
             display("Invalid cargo-release config item found")
         }
         SemVerError(err: SemVerError) {
             from()
-            cause(err)
+            source(err)
             display("SemVerError {}", err)
         }
         IgnoreError(err: ignore::Error) {
             from()
-            cause(err)
+            source(err)
             display("ignore-pattern {}", err)
         }
         Utf8Error(err: Utf8Error) {
             from()
-            cause(err)
+            source(err)
             display("Utf8Error {}", err)
         }
         FromUtf8Error(err: FromUtf8Error) {
             from()
-            cause(err)
+            source(err)
             display("FromUtf8Error {}", err)
         }
         NoPackage {
@@ -73,7 +77,7 @@ quick_error! {
         }
         ReplacerRegexError(err: RegexError) {
             from()
-            cause(err)
+            source(err)
             display("RegexError {}", err)
         }
         ReplacerMinError(pattern: String, req: usize, actual: usize) {
@@ -84,7 +88,7 @@ quick_error! {
         }
         VarError(err: VarError) {
             from()
-            cause(err)
+            source(err)
             display("Environment Variable Error: {}", err)
         }
         GitError {