+++ src/main.rs
@@ -702,6 +702,9 @@ fn release_packages<'m>(
 
             if pkg.config.registry().is_none() {
                 cargo::wait_for_publish(crate_name, &base.version_string, timeout, dry_run)?;
+                // HACK: Even once the index is updated, there seems to be another step before the publish is fully ready.
+                // We don't have a way yet to check for that, so waiting for now in hopes everything is ready
+                std::thread::sleep(std::time::Duration::from_secs(5));
             } else {
                 log::debug!("Not waiting for publish because the registry is not crates.io and doesn't get updated automatically");
             }
@@ -764,9 +767,20 @@ fn release_packages<'m>(
                 version: Some(&base.version_string),
                 crate_name: Some(crate_name),
                 date: Some(NOW.as_str()),
+                tag_name: pkg.tag.as_ref().map(|s| s.as_str()),
                 next_version: Some(updated_version_string),
                 ..Default::default()
             };
+            if !pkg.config.post_release_replacements().is_empty() {
+                // try replacing text in configured files
+                do_file_replacements(
+                    pkg.config.post_release_replacements(),
+                    &template,
+                    cwd,
+                    false, // post-release replacements should always be applied
+                    dry_run,
+                )?;
+            }
             let commit_msg = template.render(pkg.config.post_release_commit_message());
 
             if ws_config.consolidate_commits() {